requirements-dev.txt
pytest==7.3.1
pytest-cov==4.1.0
moto==4.1.11
flake8==6.0.0
black==23.3.0
mypy==1.3.0
boto3-stubs[essential]==1.26.137